requirements_deepspeech.txt
soundfile
librosa
wavio
opencv-python
scikit-image
tqdm
noisereduce
resampy
mediapipe
matplotlib
h5py
face_alignment
python_speech_features
moviepy
pydub
chumpy
av
pims
dominate
setuptools==59.5.0
configargparse
wandb
pytorch-ignite
basicsr
facexlib
basicsr>=1.4.2
facexlib>=0.2.5
lmdb
pyyaml
scipy
tb-nightly
yapf
realesrgan
# pytorch-fid
tensorflow-gpu==1.15
ninja
torchvision